convert ObaRegionsTask AsyncTask to IntentService
