+++ onebusaway-android/src/amazon/java/org/onebusaway/android/map/googlemapsv2/BaseMapFragment.java
@@ -51,7 +51,7 @@ import org.onebusaway.android.map.MapModeController;
 import org.onebusaway.android.map.MapParams;
 import org.onebusaway.android.map.RouteMapController;
 import org.onebusaway.android.map.StopMapController;
-import org.onebusaway.android.region.ObaRegionsTask;
+import org.onebusaway.android.region.ObaRegionsService.ObaRegionsReceiver;
 import org.onebusaway.android.util.LocationHelper;
 import org.onebusaway.android.util.UIHelp;
 
@@ -86,7 +86,7 @@ import java.util.List;
  * @author paulw, barbeau
  */
 public class BaseMapFragment extends SupportMapFragment
-        implements MapModeController.Callback, ObaRegionsTask.Callback,
+        implements MapModeController.Callback, ObaRegionsReceiver.Callback,
         MapModeController.ObaMapView,
         LocationSource, LocationHelper.Listener,
         com.amazon.geo.mapsv2.AmazonMap.OnCameraChangeListener,
