+++ onebusaway-android/src/google/java/org/onebusaway/android/map/googlemapsv2/BaseMapFragment.java
@@ -40,7 +40,7 @@ import org.onebusaway.android.map.MapModeController;
 import org.onebusaway.android.map.MapParams;
 import org.onebusaway.android.map.RouteMapController;
 import org.onebusaway.android.map.StopMapController;
-import org.onebusaway.android.region.ObaRegionsTask;
+import org.onebusaway.android.region.ObaRegionsService.ObaRegionsReceiver;
 import org.onebusaway.android.util.LocationHelper;
 import org.onebusaway.android.util.UIHelp;
 
@@ -75,7 +75,7 @@ import java.util.List;
  * @author paulw, barbeau
  */
 public class BaseMapFragment extends SupportMapFragment
-        implements MapModeController.Callback, ObaRegionsTask.Callback,
+        implements MapModeController.Callback, ObaRegionsReceiver.Callback,
         MapModeController.ObaMapView,
         LocationSource, LocationHelper.Listener,
         com.google.android.gms.maps.GoogleMap.OnCameraChangeListener,
