+++ onebusaway-android/src/main/AndroidManifest.xml
@@ -146,6 +146,8 @@
             </intent-filter>
         </service>
 
+        <service android:name=".region.ObaRegionsService" android:exported="false"/>
+
         <receiver android:name=".tripservice.AlarmReceiver">
             <intent-filter>
                 <!-- Should match constants for actions defined in TripService -->
