+++ onebusaway-android/src/main/java/org/onebusaway/android/ui/HomeActivity.java
@@ -22,6 +22,7 @@ import android.content.ActivityNotFoundException;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.IntentFilter;
 import android.content.SharedPreferences;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
@@ -30,6 +31,7 @@ import android.location.Location;
 import android.os.Build;
 import android.os.Bundle;
 import android.support.v4.app.FragmentManager;
+import android.support.v4.content.LocalBroadcastManager;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.ActionBarActivity;
@@ -63,7 +65,8 @@ import org.onebusaway.android.io.request.ObaArrivalInfoResponse;
 import org.onebusaway.android.map.MapModeController;
 import org.onebusaway.android.map.MapParams;
 import org.onebusaway.android.map.googlemapsv2.BaseMapFragment;
-import org.onebusaway.android.region.ObaRegionsTask;
+import org.onebusaway.android.region.ObaRegionsService;
+import org.onebusaway.android.region.ObaRegionsService.ObaRegionsReceiver;
 import org.onebusaway.android.tripservice.TripService;
 import org.onebusaway.android.util.FragmentUtils;
 import org.onebusaway.android.util.LocationUtil;
@@ -167,6 +170,8 @@ public class HomeActivity extends ActionBarActivity
 
     ObaStop mFocusedStop = null;
 
+    ObaRegionsReceiver obaRegionReceiver;
+
     /**
      * Starts the MapActivity with a particular stop focused with the center of
      * the map at a particular point.
@@ -254,6 +259,14 @@ public class HomeActivity extends ActionBarActivity
         autoShowWhatsNew();
 
         checkRegionStatus();
+    }
+
+    @Override
+    public void onDestroy() {
+        super.onDestroy();
+        if (obaRegionReceiver != null) {
+            LocalBroadcastManager.getInstance(this).unregisterReceiver(obaRegionReceiver);
+        }
     }
 
     @Override
@@ -784,10 +797,15 @@ public class HomeActivity extends ActionBarActivity
             showProgressDialog = false;
         }
 
-        //Check region status, possibly forcing a reload from server and checking proximity to current region
-        ObaRegionsTask task = new ObaRegionsTask(this, mMapFragment, forceReload,
-                showProgressDialog);
-        task.execute();
+        final String obaRegionServiceFilter = "HomeActivity_obaRegionReceiver";
+        obaRegionReceiver = new ObaRegionsReceiver(this, mMapFragment, showProgressDialog);
+        LocalBroadcastManager.getInstance(this).registerReceiver(obaRegionReceiver, new IntentFilter(
+                obaRegionServiceFilter));
+        Intent obaRegionsIntent = new Intent(this, ObaRegionsService.class);
+        obaRegionReceiver.showProgressDialog();
+        obaRegionsIntent.putExtra("mForceReload", forceReload);
+        obaRegionsIntent.putExtra("FILTER", obaRegionServiceFilter);
+        this.startService(obaRegionsIntent);
     }
 
     private void setupMyLocationButton() {
