+++ onebusaway-android/src/main/java/org/onebusaway/android/ui/PreferencesActivity.java
@@ -20,9 +20,12 @@ import org.onebusaway.android.BuildConfig;
 import org.onebusaway.android.R;
 import org.onebusaway.android.app.Application;
 import org.onebusaway.android.io.ObaAnalytics;
-import org.onebusaway.android.region.ObaRegionsTask;
+import org.onebusaway.android.region.ObaRegionsService;
+import org.onebusaway.android.region.ObaRegionsService.ObaRegionsReceiver;
 
+import android.content.Context;
 import android.content.Intent;
+import android.content.IntentFilter;
 import android.content.SharedPreferences;
 import android.net.Uri;
 import android.os.Bundle;
@@ -30,6 +33,7 @@ import android.preference.ListPreference;
 import android.preference.Preference;
 import android.preference.Preference.OnPreferenceChangeListener;
 import android.preference.PreferenceActivity;
+import android.support.v4.content.LocalBroadcastManager;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.Window;
@@ -39,7 +43,9 @@ import com.google.android.gms.analytics.GoogleAnalytics;
 
 public class PreferencesActivity extends PreferenceActivity
         implements Preference.OnPreferenceClickListener, OnPreferenceChangeListener,
-        SharedPreferences.OnSharedPreferenceChangeListener, ObaRegionsTask.Callback {
+        SharedPreferences.OnSharedPreferenceChangeListener, ObaRegionsReceiver.Callback {
+
+    private static final String OBAREGIONS_RECEIVER_FILTER = "PreferencesActivity_obaRegionsReceiver";
 
     private static final String TAG = "PreferencesActivity";
 
@@ -55,6 +61,8 @@ public class PreferencesActivity extends PreferenceActivity
     //Save initial value so we can compare to current value in onDestroy()
 
     ListPreference preferredUnits;
+
+    ObaRegionsReceiver obaRegionsReceiver;
 
     @SuppressWarnings("deprecation")
     public void onCreate(Bundle savedInstanceState) {
@@ -84,6 +92,9 @@ public class PreferencesActivity extends PreferenceActivity
                 getString(R.string.preference_key_preferred_units));
 
         settings.registerOnSharedPreferenceChangeListener(this);
+        obaRegionsReceiver = new ObaRegionsReceiver(this, this, false);
+        LocalBroadcastManager.getInstance(this).registerReceiver(obaRegionsReceiver,
+                new IntentFilter(OBAREGIONS_RECEIVER_FILTER));
     }
 
     @Override
@@ -188,6 +199,9 @@ public class PreferencesActivity extends PreferenceActivity
             }
             NavHelp.goHome(this);
         }
+        if (obaRegionsReceiver != null) {
+            LocalBroadcastManager.getInstance(this).unregisterReceiver(obaRegionsReceiver);
+        }
         super.onDestroy();
     }
 
@@ -210,9 +224,11 @@ public class PreferencesActivity extends PreferenceActivity
             which will survive orientation changes.
             */
             setProgressBarIndeterminateVisibility(true);
-            ObaRegionsTask task = new ObaRegionsTask(this, this, true, false);
-            task.execute();
-
+            Intent obaRegionsIntent = new Intent(this, ObaRegionsService.class);
+            obaRegionsReceiver.showProgressDialog();
+            obaRegionsIntent.putExtra("mForceReload", true);
+            obaRegionsIntent.putExtra("FILTER", OBAREGIONS_RECEIVER_FILTER);
+            this.startService(obaRegionsIntent);
             // Wait to change the region preference description until the task callback
             //Analytics
             if (experimentalServers) {
